+++ packages/primitiveone/src/components/footer.js
@@ -44,7 +44,7 @@ const Footer = ({ state }) => (
         </Row>
         <Row>
           <Col lg="12">
-            <p>Website design and build by <a title="Website design and build by Primitive Digital" href="https://primitivedigital.uk/">Primitive Digital</a> &copy; 2019.</p>
+            <p>Website design and build by <a title="Website design and build by Primitive Digital" href="https://primitivedigital.uk/">Primitive Digital</a> &copy; 2020.</p>
             <p>Web Hosting and Domains provided by <a title="Web Hosting and Domains from Primitive Hosting" href="https://primitivehosting.uk/">Primitive Hosting</a>.</p>
           </Col>
         </Row>