+++ packages/primitiveone/src/index.js
@@ -1,30 +1,50 @@
 import Theme from "./components";
 import image from "@frontity/html2react/processors/image";
-//import '../../../node_modules/bootstrap/dist/css/bootstrap.min.css';
-
-const before = ({ libraries }) => {
-  // We use html2react to process the <img> tags inside the content HTML.
-  libraries.html2react.processors.push(image);
-};
 
 const primitiveOne = {
   name: "primitiveone",
   roots: {
+    /** src: frontity mars-theme comments
+     *  In Frontity, any package can add React components to the site.
+     *  We use roots for that, scoped to the `theme` namespace.
+     */
     theme: Theme
   },
   state: {
+    /** src: frontity mars-theme comments
+     * State is where the packages store their default settings and other
+     * relevant state. It is scoped to the `theme` namespace.
+     */
     theme: {
       menu: [],
+      isMobileMenuOpen: false,
       featured: {
         showOnList: false,
         showOnPost: false
       }
     }
   },
+  /** src: frontity mars-theme comments
+   * Actions are functions that modify the state or deal with other parts of
+   * Frontity like libraries.
+   */
   actions: {
     theme: {
-      beforeSSR: before,
-      beforeCSR: before
+      toggleMobileMenu: ({ state }) => {
+        state.theme.isMobileMenuOpen = !state.theme.isMobileMenuOpen;
+      },
+      closeMobileMenu: ({ state }) => {
+        state.theme.isMobileMenuOpen = false;
+      }
+    }
+  },
+  libraries: {
+    html2react: {
+      /** src: frontity mars-theme comments
+       * Add a processor to `html2react` so it processes the `<img>` tags
+       * inside the content HTML. You can add your own processors too
+       */
+      processors: [image]
     }
   }
 };